Refactor form fields to use HasCustomOptions trait and update PHPStan configuration
