+++ phpstan.neon.dist
@@ -13,6 +13,8 @@ parameters:
         - '#Return type \(Illuminate\\Contracts\\View\\Factory\|Illuminate\\View\\View\) of method Encore\\Admin\\Widgets\\Grid\\Tools\\QuickSearch::render\(\) should be compatible with return type \(string\) of method Encore\\Admin\\Widgets\\Grid\\Tools\\AbstractTool::render\(\)#'
         - '#Return type \(Illuminate\\Contracts\\View\\Factory\|Illuminate\\View\\View\) of method Encore\\Admin\\Widgets\\Grid\\Tools\\QuickSearch::render\(\) should be compatible with return type \(string\) of method Illuminate\\Contracts\\Support\\Renderable::render\(\)#'
         - '#Instanceof between Illuminate\\Database\\Eloquent\\Relations\\Relation and Illuminate\\Database\\Eloquent\\Relations\\MorphToMany will always evaluate to false.#'
+        - '#Call to static method where\(\) on an unknown class Exceedone\\Exment\\Model\\File\.#'
+        - '#Access to constant AI_OCR on an unknown class Exceedone\\Exment\\Enums\\FileType\.#'
         - '#Illuminate\\Database\\Eloquent\\Relations\\BelongsToMany does not specify its types#'
         - '#Illuminate\\Database\\Eloquent\\Relations\\BelongsTo does not specify its types#'
