+++ src/Form/Field/Currency.php
@@ -4,18 +4,7 @@ namespace Encore\Admin\Form\Field;
 
 class Currency extends Text
 {
-    protected $customOptions = [];
-
-    public function setCustomOptions(array $customOptions)
-    {
-        $this->customOptions = $customOptions;
-        return $this;
-    }
-
-    public function getCustomOptions()
-    {
-        return $this->customOptions;
-    }
+    use HasCustomOptions;
 
     /**
      * @var string
