+++ src/Form/Field/Textarea.php
@@ -6,18 +6,7 @@ use Encore\Admin\Form\Field;
 
 class Textarea extends Field
 {
-    protected $customOptions = [];
-
-    public function setCustomOptions(array $customOptions)
-    {
-        $this->customOptions = $customOptions;
-        return $this;
-    }
-
-    public function getCustomOptions()
-    {
-        return $this->customOptions;
-    }
+    use HasCustomOptions;
 
     /**
      * Default rows of textarea.
