+++ src/Widgets/Form.php
@@ -3,6 +3,7 @@
 namespace Encore\Admin\Widgets;
 
 use Closure;
+use Encore\Admin\Admin;
 use Encore\Admin\Form as BaseForm;
 use Encore\Admin\Form\Field;
 use Encore\Admin\Traits\FormTrait;
@@ -711,7 +712,7 @@ $('.after-submit').iCheck({checkboxClass:'icheckbox_minimal-blue'}).on('ifChecke
 });
 EOT;
 
-        \Admin::script($script);
+        Admin::script($script);
     }
 
     /**
